Clean up brand stuff
