--- xikolo-common/Branding/XikoloBrandColor.swift
@@ -1,33 +1,0 @@
-//
-//  Created for xikolo-ios under MIT license.
-//  Copyright © HPI. All rights reserved.
-//
-
-//import UIKit
-//
-//protocol XikoloBrandColor {
-//    static var primary: UIColor { get }
-//    static var secondary: UIColor { get }
-//    static var tertiary: UIColor { get }
-//    static var window: UIColor { get }
-//}
-//
-//extension XikoloBrandColor {
-//
-//    static var window: UIColor {
-//        return Self.primary
-//    }
-//
-//    static var answerCorrect: UIColor {
-//        return UIColor(red: 140 / 255, green: 179 / 255, blue: 13 / 255, alpha: 1)
-//    }
-//
-//    static var answerIncorrect: UIColor {
-//        return UIColor(red: 214 / 255, green: 0 / 255, blue: 26 / 255, alpha: 1)
-//    }
-//
-//    static var answerWrong: UIColor {
-//        return UIColor(red: 187 / 255, green: 188 / 255, blue: 190 / 255, alpha: 1)
-//    }
-//
-//}
