--- xikolo-common/Branding/openHPI/Brand.swift
@@ -1,25 +1,0 @@
-//
-//  Created for xikolo-ios under MIT license.
-//  Copyright © HPI. All rights reserved.
-//
-
-//import Foundation
-//import UIKit
-//
-//struct Brand: XikoloBrand {
-//
-//    struct Color: XikoloBrandColor {
-//        static let primary = UIColor(red: 222 / 255, green: 98 / 255, blue: 18 / 255, alpha: 1.0)
-//        static let secondary = UIColor(red: 180 / 255, green: 41 / 255, blue: 70 / 255, alpha: 1.0)
-//        static let tertiary = UIColor(red: 245 / 255, green: 167 / 255, blue: 4 / 255, alpha: 1.0)
-//    }
-//
-//    static let host = "open.hpi.de"
-//    static let imprintURL = Routes.base.appendingPathComponents(["pages", "imprint"])
-//    static let privacyURL = Routes.base.appendingPathComponents(["pages", "privacy"])
-//
-//    static let platformTitle = "hpi"
-//
-//    static let copyrightName = "HPI"
-//
-//}
